+++ src/main/java/com/gms/web/admin/domain/manage/CustomerPriceExtVO.java
@@ -21,10 +21,10 @@ public class CustomerPriceExtVO implements Serializable {
 	private Integer productPriceSeq;	
 	
 	/** Product_Price 	*/    
-	private int productPrice;
+	private double productPrice;
 	
 	/** Product_Bottle_Price 	*/    
-	private int productBottlePrice;
+	private double productBottlePrice;
 	
 	/**Product_Nm */
 	private String productNm;
