+++ src/main/java/com/gms/web/admin/domain/statistics/StatisticsUserVO.java
@@ -23,13 +23,13 @@ public class StatisticsUserVO extends AbstractVO implements Serializable {
 	private int salesCount;
 	
 	/**Sales_Amount    */
-	private int salesAmount;
+	private double salesAmount;
 
 	/**Rental_Count     */
 	private int rentalCount;	
 	
 	/**Rental_Amount    */
-	private int rentalAmount;
+	private double rentalAmount;
 	
 	/** searchStatDt	*/
 	private String  searchStatDt;
